refactor: rename utils to constant
